extract to yearly recurrence module
